# infer: make `cset-meet` lazy