Use a stream to avoid eagerly computing `cset-meet`.

Good news:
- quickly passes the unit test in #743 
- fixes the slowdown for the mutable/immutable vectors PR (see #575 )

Bad news:
- does not fix the _program_ in #743 --- I'll make a note in that issue and work on that program after mutable/immutable vectors are done
- makes typechecking slower overall on the `tr-performance` programs:

```
Compile time ratio (old time = Racket 7.0 release / new time = Racket 7.0.0.10):
schml-specify-rep: 0.97 (i.e. slower, 7.95s (σ 0.06) to 8.21s (σ 0.03))
schml-interp-casts-help: 0.99 (i.e. slower, 17.91s (σ 0.2) to 18.12s (σ 0.11))
parser: 0.92 (i.e. slower, 2.01s (σ 0.01) to 2.19s (σ 0.01))
old-metrics: 0.92 (i.e. slower, 2.37s (σ 0.01) to 2.57s (σ 0.02))
new-metrics: 0.91 (i.e. slower, 3.21s (σ 0.01) to 3.54s (σ 0.05))
math-flonum: 0.95 (i.e. slower, 3.42s (σ 0.02) to 3.59s (σ 0.02))
fsm: 0.91 (i.e. slower, 4.51s (σ 0.02) to 4.96s (σ 0.03))
forth: 0.9 (i.e. slower, 4.36s (σ 0.04) to 4.87s (σ 0.02))
dungeon: 0.91 (i.e. slower, 7.81s (σ 0.06) to 8.54s (σ 0.05))
bernoulli: 0.91 (i.e. slower, 3.59s (σ 0.03) to 3.95s (σ 0.06))
acquire: 0.91 (i.e. slower, 14.14s (σ 0.09) to 15.52s (σ 0.07))
```

Full data (12 runs per file): 
[issue-743-lazy-perf.txt](https://github.com/racket/typed-racket/files/2263382/issue-743-lazy-perf.txt)


This PR is labeled "in-progress" because I think the `tr-performance` could be better in the common case. Maybe by using a generator instead of `for*/stream`, maybe by using lists normally and streams only when we know inference needs laziness.

- - - 

Edit: I tried debugging the performance today, and it turns out the current Racket+TR master has the same slowdown relative to Racket 7

```
Compile time ratio (old time = Racket 7 release / new time = Racket 7.0.0.10 @ 8deaa4cf91b with TR @ a3871e685a):
schml-specify-rep: 0.98 (i.e. slower, 7.94s (σ 0.09) to 8.09s (σ 0.1))
schml-interp-casts-help: 0.98 (i.e. slower, 17.81s (σ 0.17) to 18.19s (σ 0.14))
parser: 0.92 (i.e. slower, 1.98s (σ 0.02) to 2.16s (σ 0.01))
old-metrics: 0.93 (i.e. slower, 2.35s (σ 0.01) to 2.54s (σ 0.03))
new-metrics: 0.95 (i.e. slower, 3.2s (σ 0.02) to 3.38s (σ 0.02))
math-flonum: 0.95 (i.e. slower, 3.39s (σ 0.03) to 3.55s (σ 0.03))
fsm: 0.91 (i.e. slower, 4.45s (σ 0.03) to 4.9s (σ 0.05))
forth: 0.9 (i.e. slower, 4.29s (σ 0.02) to 4.77s (σ 0.04))
dungeon: 0.91 (i.e. slower, 7.72s (σ 0.06) to 8.48s (σ 0.06))
bernoulli: 0.88 (i.e. slower, 3.54s (σ 0.03) to 4.04s (σ 0.05))
acquire: 0.92 (i.e. slower, 14.03s (σ 0.13) to 15.29s (σ 0.15))
```

Full data: 
[racket7-vs-master.txt](https://github.com/racket/typed-racket/files/2279530/racket7-vs-master.txt)


This PR might still have a performance issue, but the two datasets above are not proof of that.