@pnwamk 2018-08-07
I see that we’re using ‘in-dict’ currently. Does changing that to
‘in-assoc’ (as opposed to ‘in-stream’) make a difference for the program
you were looking to help? (I ask because the ‘dict’ interface has been a
source of significant overhead before, so I can’t help but be suspicious).
On Mon, Aug 6, 2018 at 7:41 PM Ben Greenman <notifications@github.com>
wrote:

> Use a stream to avoid eagerly computing cset-meet.
>
> Good news:
>
>    - quickly passes the unit test in #743
>    <https://github.com/racket/typed-racket/issues/743>
>    - fixes the slowdown for the mutable/immutable vectors PR (see #575
>    <https://github.com/racket/typed-racket/pull/575> )
>
> Bad news:
>
>    - does not fix the *program* in #743
>    <https://github.com/racket/typed-racket/issues/743> --- I'll make a
>    note in that issue and work on that program after mutable/immutable vectors
>    are done
>    - makes typechecking slower overall on the tr-performance programs:
>
> Compile time ratio (old time = Racket 7.0 release / new time = Racket 7.0.0.10):
> schml-specify-rep: 0.97 (i.e. slower, 7.95s (σ 0.06) to 8.21s (σ 0.03))
> schml-interp-casts-help: 0.99 (i.e. slower, 17.91s (σ 0.2) to 18.12s (σ 0.11))
> parser: 0.92 (i.e. slower, 2.01s (σ 0.01) to 2.19s (σ 0.01))
> old-metrics: 0.92 (i.e. slower, 2.37s (σ 0.01) to 2.57s (σ 0.02))
> new-metrics: 0.91 (i.e. slower, 3.21s (σ 0.01) to 3.54s (σ 0.05))
> math-flonum: 0.95 (i.e. slower, 3.42s (σ 0.02) to 3.59s (σ 0.02))
> fsm: 0.91 (i.e. slower, 4.51s (σ 0.02) to 4.96s (σ 0.03))
> forth: 0.9 (i.e. slower, 4.36s (σ 0.04) to 4.87s (σ 0.02))
> dungeon: 0.91 (i.e. slower, 7.81s (σ 0.06) to 8.54s (σ 0.05))
> bernoulli: 0.91 (i.e. slower, 3.59s (σ 0.03) to 3.95s (σ 0.06))
> acquire: 0.91 (i.e. slower, 14.14s (σ 0.09) to 15.52s (σ 0.07))
>
> Full data (12 runs per file):
> issue-743-lazy-perf.txt
> <https://github.com/racket/typed-racket/files/2263382/issue-743-lazy-perf.txt>
>
> This PR is labeled "in-progress" because I think the tr-performance could
> be better in the common case. Maybe by using a generator instead of
> for*/stream, maybe by using lists normally and streams only when we know
> inference needs laziness.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/racket/typed-racket/pull/758
> Commit Summary
>
>    - infer: use stream
>
> File Changes
>
>    - *M* typed-racket-lib/typed-racket/infer/constraint-structs.rkt
>    <https://github.com/racket/typed-racket/pull/758/files#diff-0> (6)
>    - *M* typed-racket-lib/typed-racket/infer/constraints.rkt
>    <https://github.com/racket/typed-racket/pull/758/files#diff-1> (31)
>    - *M* typed-racket-lib/typed-racket/infer/infer-unit.rkt
>    <https://github.com/racket/typed-racket/pull/758/files#diff-2> (9)
>    - *M* typed-racket-test/unit-tests/infer-tests.rkt
>    <https://github.com/racket/typed-racket/pull/758/files#diff-3> (7)
>
> Patch Links:
>
>    - https://github.com/racket/typed-racket/pull/758.patch
>    - https://github.com/racket/typed-racket/pull/758.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/racket/typed-racket/pull/758>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADfg6obWlZSGuW6e4Fk7eeiRWh-vFO4iks5uONREgaJpZM4VxQue>
> .
>