@bennn 2018-08-11
Ok, the slowdown I posted above seems to be an issue with the 7.0 release vs. the current Racket and Typed Racket master --- unrelated to this pull request. Removing the in-progress label.

Here's the result of a fair comparison with a recent TR master vs. this pull request:

```
Compile time ratio (old time = Racket @ 8deaa4c TR @ 3871e68 / new time = Racket @ 8deaa4c TR @ 6a247d7e):
schml-specify-rep: 0.99 (i.e. slower, 8.09s (σ 0.08) to 8.16s (σ 0.08))
schml-interp-casts-help: 1.02 (i.e. faster, 18.04s (σ 0.34) to 17.72s (σ 0.24))
parser: 1.0 (i.e. faster, 2.15s (σ 0.02) to 2.14s (σ 0.02))
old-metrics: 1.0 (i.e. slower, 2.52s (σ 0.02) to 2.52s (σ 0.02))
new-metrics: 0.98 (i.e. slower, 3.35s (σ 0.02) to 3.42s (σ 0.02))
math-flonum: 1.01 (i.e. faster, 3.56s (σ 0.02) to 3.54s (σ 0.02))
fsm: 1.0 (i.e. faster, 4.88s (σ 0.04) to 4.86s (σ 0.05))
forth: 0.99 (i.e. slower, 4.72s (σ 0.03) to 4.75s (σ 0.04))
dungeon: 1.01 (i.e. faster, 8.41s (σ 0.07) to 8.36s (σ 0.05))
bernoulli: 1.06 (i.e. faster, 4.03s (σ 0.04) to 3.8s (σ 0.03))
acquire: 0.99 (i.e. slower, 15.08s (σ 0.11) to 15.28s (σ 0.1))
```
Full data (15 runs): 
[orig-vs-master.log](https://github.com/racket/typed-racket/files/2279806/orig-vs-master.log)